Implement basic job management for manager and employee
